requirements.txt
openai
PyPDF2
langchain
langchain-community
langchain-openai
python-dotenv
supabase
